Add structured data for models
